% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cells_from_sheet.R
\name{extract_cells_from_sheet}
\alias{extract_cells_from_sheet}
\title{Read every cell in a single sheet and save them with their addresses}
\usage{
extract_cells_from_sheet(.spreadsheet, .sheet)
}
\arguments{
\item{.spreadsheet}{the full path to the workbook file}

\item{.sheet}{the name or number of the sheet to parse}
}
\value{
a tibble with 3 columns
\describe{
\item{Row}{\if{html}{\out{<int>}} The row number of the cell}
\item{Column}{\if{html}{\out{<int>}} The column number of the cell}
\item{Value}{\if{html}{\out{<chr>}} The literal contents of the cell}
}
}
\description{
Read every cell in a single sheet and save them with their addresses
}
