% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cells_from_spreadsheet.R
\name{extract_cells_from_spreadsheet}
\alias{extract_cells_from_spreadsheet}
\title{Read every cell in an entire spreadsheet and save them with their addresses}
\usage{
extract_cells_from_spreadsheet(.spreadsheet)
}
\arguments{
\item{.spreadsheet}{the full path to the workbook file}
}
\value{
a tibble with 4 columns
\describe{
\item{Sheet}{\if{html}{\out{<chr>}} The sheet that the cell came from}
\item{Row}{\if{html}{\out{<int>}} The row number of the cell}
\item{Column}{\if{html}{\out{<int>}} The column number of the cell}
\item{Value}{\if{html}{\out{<chr>}} The literal contents of the cell}
}
}
\description{
Read every cell in an entire spreadsheet and save them with their addresses
}
