% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smeltr-package.R
\docType{package}
\name{smeltr-package}
\alias{smeltr}
\alias{smeltr-package}
\title{smeltr: State Machines for Extracting the Layout of Tables in R}
\description{
Build algorithms for tidying spreadsheets with complicated layouts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/smeltr/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
