% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STOP_SHEETS}
\alias{STOP_SHEETS}
\title{A character vector of sheets that won't have tabular data}
\format{
An object of class \code{character} of length 3.
}
\usage{
STOP_SHEETS
}
\description{
A character vector of sheets that won't have tabular data
}
\keyword{datasets}
