% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cells_from_sheet.R
\name{extract_cells_from_sheet}
\alias{extract_cells_from_sheet}
\title{Read every cell in a single sheet and save them with their addresses}
\usage{
extract_cells_from_sheet(.spreadsheet, .sheet)
}
\arguments{
\item{.spreadsheet}{\verb{<chr>} the full path to the workbook file}

\item{.sheet}{\verb{<chr|int>} the name or number of the sheet to parse}
}
\value{
a tibble with 3 columns
\describe{
\item{Row}{\verb{<int>} The row number of the cell}
\item{Column}{\verb{<int>} The column number of the cell}
\item{Value}{\verb{<chr>} The literal contents of the cell}
}
}
\description{
Read every cell in a single sheet and save them with their addresses
}
