% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cells_from_spreadsheet.R
\name{extract_cells_from_spreadsheet}
\alias{extract_cells_from_spreadsheet}
\title{Read every cell in an entire spreadsheet and save them with their addresses}
\usage{
extract_cells_from_spreadsheet(.spreadsheet)
}
\arguments{
\item{.spreadsheet}{\verb{<chr>} the full path to the workbook file}
}
\value{
a tibble with 4 columns
\describe{
\item{Sheet}{\verb{<chr>} The sheet that the cell came from}
\item{Row}{\verb{<int>} The row number of the cell}
\item{Column}{\verb{<int>} The column number of the cell}
\item{Value}{\verb{<chr>} The literal contents of the cell}
}
}
\description{
Read every cell in an entire spreadsheet and save them with their addresses
}
