% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_headers_from_sheet.R
\name{extract_headers_from_sheet}
\alias{extract_headers_from_sheet}
\title{Read header information from one sheet in a workbook}
\usage{
extract_headers_from_sheet(.spreadsheet, .sheet, .nrows)
}
\arguments{
\item{.spreadsheet}{the full path to the workbook file}

\item{.sheet}{the name or number of the sheet to parse}

\item{.nrows}{the number of rows to investigate as headers}
}
\value{
a tibble with 1 + \code{.nrows} columns.
\describe{
\item{Column}{\if{html}{\out{<int>}} which column the value came from}
\item{Row[i]}{\if{html}{\out{<chr>}} the header from row \code{i} in this Column.}
}
}
\description{
Read header information from one sheet in a workbook
}
